Revert last commit __BYPASS_CHECKS__
git-svn-id: svn+ssh://arcadia.yandex.ru/arc/trunk/arcadia/yt/19_2/yt@3217780 41d65440-b5be-11dd-afe3-b2e846d9b4f8

+++ server/job_agent/config.h
@@ -16,7 +16,7 @@ class TResourceLimitsConfig
 {
 public:
     int UserSlots;
-    double Cpu;
+    int Cpu;
     int Network;
     i64 Memory;
     int ReplicationSlots;
